pyproject.toml
[project]
name = "hongyuan-sr"
version = "0.1.0"
description = "District 8 SR Private Server"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "betterproto2-compiler>=0.9.0",
    "betterproto2[all]>=0.9.1",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
]
